Moved subtitle timeline code to own file
